Delete 맛탱이 가있는데 혹시몰라서 일단 push

+++ Test/src/main/java/controller/FreeDelete.java
@@ -6,7 +6,9 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
+import dto.FBoard;
 import service.FreeService;
 import service.FreeServiceImpl;
 import service.TempService;
@@ -33,13 +35,38 @@ public class FreeDelete extends HttpServlet {
 	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		request.setCharacterEncoding("utf-8");
 		Integer freeNum = Integer.parseInt(request.getParameter("freeNum"));
-		try {
-			FreeService freeService=new FreeServiceImpl();
-			freeService.deleteBoard(freeNum);
-			response.sendRedirect("freeboard");
-		} catch(Exception e) {
-			e.printStackTrace();
-		}
+//		try {
+//			FreeService freeService=new FreeServiceImpl();
+//			freeService.deleteBoard(freeNum);
+//			response.sendRedirect("freeboard");
+//		} catch(Exception e) {
+//			e.printStackTrace();
+//		}
+		 HttpSession session = request.getSession(false);
+		 if (session != null && session.getAttribute("user") != null) {
+			 try {
+				 String memNick = (String) session.getAttribute("user");
+				 FreeService freeService=new FreeServiceImpl();
+				 FBoard board = freeService.getBoardByFreeNum(freeNum);
+	            if (board != null && board.getFreeNick().equals(memNick)) {
+                    freeService.deleteBoard(freeNum);
+                    response.sendRedirect("freeboard");
+                } else {
+                    // 게시물 작성자와 로그인한 사용자가 다른 경우
+                    // 에러 처리 또는 권한 없음 페이지로 리다이렉트
+                	response.sendRedirect("/MoongStar/login");
+                }
+            } catch(Exception e) {
+                e.printStackTrace();
+            }
+        } else {
+            // 로그인하지 않은 경우
+            // 로그인 페이지로 리다이렉트
+            response.sendRedirect("/MoongStar/login");
+        }
+
+	        
+	                
 	}
 
 	/**
